+++ src/test/java/roomescape/theme/application/ThemeCommandUseCaseImplTest.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.theme.application.dto.CreateThemeServiceRequest;
+import roomescape.theme.application.usecase.ThemeCommandUseCaseImpl;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.domain.ThemeDescription;
 import roomescape.theme.domain.ThemeId;
